Change folder names so new Epic lives in Epic
